requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
scikit-surprise==1.1.3
python-dotenv==1.0.0
pydantic==2.5.2
joblib==1.3.2
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.3.0
pytest==7.4.3